# rocBLAS: define highlevel dot, gemm, axpy functions for FP16